4.2 Sending Rays Into the Scene
I had some trouble with the service worker. It doesn't throw an error
in the console, instead it fires an error event that you need to catch
in the main thread.

The error was an import problem, maybe because of that it didn't have a
message in the main thread.

Also had an issue with forgetting to scale the vectors to 255.